requirements.txt
attrs==25.3.0
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
h11==0.16.0
idna==3.10
numpy==2.2.6
opencv-python==4.12.0.88
outcome==1.3.0.post0
packaging==25.0
pillow==11.3.0
pycparser==2.23
PySocks==1.7.1
pytesseract==0.3.13
python-dotenv==1.1.1
requests==2.32.5
selenium==4.35.0
sniffio==1.3.1
sortedcontainers==2.4.0
trio==0.30.0
trio-websocket==0.12.2
typing_extensions==4.14.1
urllib3==2.5.0
webdriver-manager==4.0.2
websocket-client==1.8.0
wsproto==1.2.0
